requirements.txt
apscheduler>=3.10,<4
requests>=2.31,<3
python-dotenv>=1.0,<2
pyyaml>=6.0,<7
pytest>=8.0,<9
